@@ -1,2 +1,2 @@
-export const prerender = true;
+export const prerender = false;
 export const ssr = false;
