@@ -1,30 +1,15 @@
 <script>
-    import { page } from "$lib/store.js";
-
-    $page = "/";
+    import Banner from "$lib/components/Banner.svelte";
 </script>
 
 <svelte:head>
     <title>Matt Collecutt</title>
+    <meta name="description" content="Home page" />
 </svelte:head>
 
-<h1>Matt Collecutt</h1>
 
-<article class="pt-12 flex flex-col gap-4">
-    <p>
-        Hi, I’m Matt! I’m passionate about a lot of things. Here you’ll find observations
-        I’ve made about different topics with a focus around software development.
-    </p>
+<Banner></Banner>
 
-    <p>
-        The main reason for starting this blog is to improve my written communication skills,
-        it’s always been somewhat of a struggle. I hope you’ll enjoy what I have to offer in
-        the form of blog posts and that you’ll leave having learnt something new.
-    </p>
-
-    <p>
-        I’m still exploring the direction I want to take this website, so the content may evolve
-        over time. I’ve love to hear your thoughts and feedback as I continue to develop the site,
-        so feel free to leave a comment or get in touch.
-    </p>
-</article>
+<div class="mx-auto w-full max-w-3xl py-12">
+    content
+</div>
